Update README.md
This is a project written in C language for one of my CS courses. It's a classic tic-tac-toe game playable directly in the terminal.
How to Play:
Compile the game.c using GCC or any compiler of your choice, then run it in your terminal (Linux and Mac) or command prompt (Windows).
Features
Playable in the terminal for a nostalgic 
Implements the classic rules of tic-tac-toe.
Fun for all ages.

Issues Faced and Solutions

Included in this repository is a list of issues faced during development and the solutions used to solve them. Feel free to check them out!

Cheers!